fix counter resetting when form invalid
