--- src/components/Signup/ConfirmForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ConfirmForm";
